requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
cx-Oracle==8.3.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==18.4.0
gunicorn==20.1.0
iniconfig==2.0.0
numpy==1.23.3
oracledb==1.3.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pytest==7.3.1
pytest-django==4.5.2
pytest-faker==2.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.4.0
django-cors-headers==3.11.0
